greenRepo/requirements.txt
amqp==2.4.0
billiard==3.5.0.5
celery==4.2.1
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-dbbackup==3.2.0
django-enumfields==0.10.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
gunicorn==19.7.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
kombu==4.2.2.post1
MarkupSafe==1.1.1
mysqlclient==1.3.13
oauthlib==3.1.0
openapi-codec==1.3.2
python3-openid==3.1.0
pytz==2018.4
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
vine==1.2.0
whitenoise==3.3.1
